v_0.0.3 #35 소스 분리
1. 소스코드 분리
