+++ src/js/configs/attrs/default_id.js
@@ -1,0 +1,15 @@
+/**
+    Default id
+**/
+
+module.exports = {
+  ID: 'hb_attr_id',
+  TYPE: 'hb_attr_set_type', //text, text-unit...
+  ATTR_TYPE: 'hb_attr_type', //style, attr, exception
+  ATTR_TYPE_OPTION : {
+    units: 'units',
+    select: 'select',
+    text: 'text',
+    color: 'color'
+  }
+}
